Slutt med dobbelinnhenting/sammenligning Sigrun
