+++ pom.xml
@@ -34,10 +34,10 @@
 
         <kontrakt.java.version>17</kontrakt.java.version>
 
-		<felles.version>4.1.43</felles.version>
+		<felles.version>4.1.51</felles.version>
 		<prosesstask.version>3.1.10</prosesstask.version>
         <maven.deploy.skip>true</maven.deploy.skip>
-		<kontrakter.version>6.1.25</kontrakter.version>
+		<kontrakter.version>6.1.27</kontrakter.version>
 
 		<tjenestespesifikasjoner.version>1.2021.02.22-10.45-4201aaea72fb</tjenestespesifikasjoner.version>
 	</properties>
